soc-simulator/simulator/requirements.txt
redis==5.0.1
faker==20.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
